Update mainDialog test, remove the updates of the sample skill
